Run db setup only once.

+++ recipes/setup_db.rb
@@ -3,6 +3,16 @@ execute "delete db" do
   group "gitlab"
   cwd "/home/gitlab/gitlab"
   command "bundle exec rake db:drop RAILS_ENV=production"
+  notifies :create, "ruby_block[db_setup]", :immediately
+  not_if { node.attribute?("db_setup") }
+end
+
+ruby_block "db_setup" do
+  block do
+    node.set['db_setup'] = true
+    node.save
+  end
+  action :nothing
 end
 
 execute "create db and load schema" do
@@ -10,6 +20,7 @@ execute "create db and load schema" do
   group "gitlab"
   cwd "/home/gitlab/gitlab"
   command "bundle exec rake db:setup RAILS_ENV=production"
+  not_if { node.attribute?("db_setup") }
 end
 
 execute "seed db" do
@@ -17,6 +28,7 @@ execute "seed db" do
   group "gitlab"
   cwd "/home/gitlab/gitlab"
   command "bundle exec rake db:seed_fu RAILS_ENV=production"
+  not_if { node.attribute?("db_setup") }
 end
 
 execute "run status" do
